Extract tabs
refs: MBL-18907
affects: Teacher
release note: none
