+++ Teacher/Teacher/SpeedGrader/SpeedGraderPage/ViewModel/SpeedGraderPageTab.swift
@@ -1,0 +1,31 @@
+//
+// This file is part of Canvas.
+// Copyright (C) 2025-present  Instructure, Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as
+// published by the Free Software Foundation, either version 3 of the
+// License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+//
+
+import Foundation
+
+enum SpeedGraderPageTab: Int, CaseIterable {
+    case grades
+    case comments
+
+    var title: String {
+        switch self {
+        case .grades: String(localized: "Grades", bundle: .teacher)
+        case .comments: String(localized: "Comments", bundle: .teacher)
+        }
+    }
+}
